+++ cardano-api/internal/Cardano/Api/Eras.hs
@@ -49,5 +49,4 @@ module Cardano.Api.Eras
   ) where
 
 import           Cardano.Api.Eras.Case
-import           Cardano.Api.Eras.Constraints
 import           Cardano.Api.Eras.Core
